Bump up Groonga from 14.0.2 to 14.0.4 (#45)
ref: https://github.com/pgroonga/docker/issues/40

This commit cannot resolve the current CI problem.
We will add the additional changes at the following PRs. Please see also
https://github.com/pgroonga/docker/issues/40#issuecomment-2190383938.

## Issue
Groonga v14.0.2 cannot find the `msgpack-c` package during the CMake
configuration step.
This issue is due to the absence of changes from the commit
[2f01dcce015a6d7cd7f312313d92fd3a5643a4c3](https://github.com/groonga/groonga/commit/2f01dcce015a6d7cd7f312313d92fd3a5643a4c3)
in the version we are using.

## Soulution
Update Groonga to v14.0.4 which includes the necessary commit to resolve
the msgpack-c package finding issue.