Improve copying of share objects
Groonga is now built by CMake. (GH-46)
More build options and more share objects to depend on.
Since it is not maintainable to `COPY` them individually,
copy the share objects we need from the `ldd` results.

Since there are many Dockerfile, only `alpine/12-slim` is updated in this commit.
The other Dockerfile will be updated in another commit.